Refactor: Decompose prompts, apiHandlers, mcrService, and evaluator
- Decomposed `src/prompts.js` into individual files in `prompts/` and updated `src/prompts.js` to load them dynamically.
- Decomposed `src/apiHandlers.js` into smaller, focused handler files in `src/api/`. Updated `src/routes.js` accordingly.
- Extracted translation and explanation logic from `src/mcrService.js` into `src/services/translationService.js`.
- Extracted metrics calculation logic from `src/evaluator.js` into `src/evaluation/metrics.js`.
- Removed API reference section from README.md as per user request.
- Added missing `EMPTY_EXPLANATION_GENERATED` error code to `src/errors.js`.
- Fixed path resolution for the `prompts/` directory in `src/prompts.js`.
- Installed `@babel/preset-env` dev dependency to resolve test issues.
