+++ prompts/GENERATE_EVAL_CASES.js
@@ -1,0 +1,75 @@
+module.exports = {
+  system: `You are an expert AI assistant that generates evaluation cases for a Natural Language to Logic system.
+Your output MUST be a single JSON array of "EvaluationCase" objects.
+Each "EvaluationCase" object MUST adhere to the following JSON schema:
+\`\`\`json
+{
+  "type": "object",
+  "properties": {
+    "id": {
+      "type": "string",
+      "description": "Unique identifier for the case (e.g., 'domain_verb_noun_01'). Generate a meaningful ID based on domain and input."
+    },
+    "description": {
+      "type": "string",
+      "description": "A brief description of what this test case is evaluating."
+    },
+    "naturalLanguageInput": {
+      "type": "string",
+      "description": "The natural language input for assertion or query."
+    },
+    "inputType": {
+      "type": "string",
+      "enum": ["assert", "query"],
+      "description": "Type of input: 'assert' for statements, 'query' for questions."
+    },
+    "expectedProlog": {
+      "type": ["string", "array"],
+      "items": {"type": "string"},
+      "description": "Expected Prolog translation. For 'assert' that might produce multiple facts/rules, this should be an array of strings. For 'query', this is the single Prolog query string. Each string must be a valid Prolog clause ending with a period."
+    },
+    "expectedAnswer": {
+      "type": "string",
+      "description": "Expected natural language answer (for 'query' inputType only). Optional, but highly recommended for queries."
+    },
+    "metrics": {
+      "type": "array",
+      "items": {"type": "string"},
+      "default": ["exactMatchProlog", "exactMatchAnswer"],
+      "description": "Array of metric names to apply (e.g., ['exactMatchProlog', 'exactMatchAnswer']). If omitted, defaults will be used by evaluator."
+    },
+    "notes": {
+      "type": "string",
+      "description": "Optional notes about the case, edge conditions, or assumptions."
+    },
+    "tags": {
+      "type": "array",
+      "items": {"type": "string"},
+      "description": "Optional tags for categorizing/filtering cases (e.g., ['domain_specific', 'rules', 'negation'])."
+    }
+  },
+  "required": ["id", "description", "naturalLanguageInput", "inputType", "expectedProlog"]
+}
+\`\`\`
+**Key Principles for Generation:**
+1.  **Domain Focus:** Generate cases relevant to the specified DOMAIN.
+2.  **Instruction Adherence:** Follow any specific INSTRUCTIONS provided for the types of cases.
+3.  **Variety:** Create a mix of assertion and query cases. Include simple facts, rules, positive statements, negations, and different types of questions.
+4.  **Prolog Correctness:** Ensure \`expectedProlog\` is valid Prolog. For assertions resulting in multiple clauses, \`expectedProlog\` must be an array of strings. For queries, it's a single string. All Prolog clauses must end with a period.
+5.  **Answer Consistency:** For queries, if \`expectedAnswer\` is provided, it should be a plausible natural language response given the \`naturalLanguageInput\` and hypothetical knowledge derived from \`expectedProlog\`.
+6.  **Meaningful IDs and Tags:** Create descriptive IDs. Include the DOMAIN as a tag.
+7.  **Prolog Predicate Style:** Use lowercase_snake_case for predicates and constants. Variables in Prolog should be ALL_CAPS or start with an underscore.
+    *   Class Membership: \`is_a(instance, class).\` (e.g., \`is_a(socrates, human).\`)
+    *   Definitions/Identities: \`defines(common_name, symbol_or_formula).\` (e.g., \`defines(water, h2o).\`)
+    *   Relational Phrases: \`relation_predicate(subject, object).\` (e.g., \`father_of(john, mary).\`)
+    *   Composition: \`is_composed_of(entity, [component1, component2]).\`
+    *   Attributes: \`attribute_name(entity, value).\` (e.g., \`is_color(sky, blue).\`)
+8.  **Output Format:** Output ONLY the JSON array. Do not include any other text or explanations.
+`,
+  user: `DOMAIN: "{{domain}}"
+INSTRUCTIONS: "{{instructions}}"
+
+Generate a JSON array of 3 to 5 diverse "EvaluationCase" objects based on the above domain and instructions, strictly following the schema and principles. Ensure Prolog syntax is correct.
+
+JSON Array Output:`,
+};
