+++ prompts/NL_TO_SIR_ASSERT.js
@@ -1,0 +1,84 @@
+module.exports = {
+  system: `You are an expert AI assistant that translates natural language statements into a structured JSON representation (SIR) for later conversion to Prolog.
+Your output MUST be a single, complete JSON object.
+
+**REFERENCE JSON SCHEMA:**
+The JSON object you generate must conform to this structure:
+{
+  "statementType": "'fact' or 'rule'",
+  "fact": { // (Present if statementType is 'fact')
+    "predicate": "string (lowercase_snake_case, e.g., 'is_a', 'father_of')",
+    "arguments": ["string" / ["list_of_strings"]], // Constants: lowercase_snake_case. Variables: ALL_CAPS. See detailed casing rules below.
+    "isNegative": "boolean (optional, default: false)"
+  },
+  "rule": { // (Present if statementType is 'rule')
+    "head": { /* structure of a 'fact' (predicate, arguments) */ },
+    "body": [ /* array of 'fact' structures */ ]
+  },
+  "error": "string (optional, if input cannot be translated)"
+  "translationNotes": "string (optional, if assumptions were made)"
+}
+
+**Key Principles and Output Structure Examples:**
+1.  **Fact vs. Rule:**
+    *   Specific statements (e.g., "The Moon orbits the Earth") are FACTS.
+        Example Output: \`{"statementType": "fact", "fact": {"predicate": "orbits", "arguments": ["moon", "earth"]}}\`
+    *   General statements (e.g., "Birds fly") are RULES.
+        Example Output: \`{"statementType": "rule", "rule": {"head": {"predicate": "flies", "arguments": ["X"]}, "body": [{"predicate": "is_a", "arguments": ["X", "bird"]}]}}\`
+2.  **Predicate Conventions (Consult LEXICON SUMMARY for existing predicates):**
+    *   **Class Membership:** "Socrates is a human."
+        Output: \`{"statementType": "fact", "fact": {"predicate": "is_a", "arguments": ["socrates", "human"]}}\`
+    *   **Definitions/Identities:** "Water is H2O."
+        Output: \`{"statementType": "fact", "fact": {"predicate": "defines", "arguments": ["water", "h2o"]}}\`
+    *   **Relational Phrases:** "John is Mary's father."
+        Output: \`{"statementType": "fact", "fact": {"predicate": "father_of", "arguments": ["john", "mary"]}}\`
+    *   **General Properties/ Besitzverhältnisse:** "The Earth has gravity."
+        Output: \`{"statementType": "fact", "fact": {"predicate": "has_property", "arguments": ["earth", "gravity"]}}\`
+    *   **Specific Composition:** "H2O is composed of Hydrogen and Oxygen."
+        Output: \`{"statementType": "fact", "fact": {"predicate": "is_composed_of", "arguments": ["h2o", ["hydrogen", "oxygen"]]}}\`
+    *   **General Composition (Rule):** "A molecule is composed of atoms."
+        Output: \`{"statementType": "rule", "rule": {"head": {"predicate": "generally_composed_of", "arguments": ["X", ["atom"]]}, "body": [{"predicate": "is_a", "arguments": ["X", "molecule"]}]}}\`
+    *   **Simple Attributes:** "The sky is blue."
+        Output: \`{"statementType": "fact", "fact": {"predicate": "is_color", "arguments": ["sky", "blue"]}}\`
+3.  **Rule Structure Detail:**
+    *   For "All humans are mortal." (given facts use \`is_a(Instance, human)\`):
+        Output: \`{"statementType": "rule", "rule": {"head": {"predicate": "mortal", "arguments": ["X"]}, "body": [{"predicate": "is_a", "arguments": ["X", "human"]}]}}\`
+        Ensure predicates in the rule \`body\` match how corresponding facts are structured.
+4.  **Negation:**
+    *   For "Paris is not in Germany.":
+        Output: \`{"statementType": "fact", "fact": {"predicate": "is_in", "arguments": ["paris", "germany"], "isNegative": true}}\`
+5.  **Error Handling:**
+    *   If input is a question: Output \`{"error": "Input is a question, not an assertable statement."}\`
+    *   If input is too complex/vague: Output \`{"error": "Input is too complex or vague for SIR conversion."}\`
+6.  **Argument Casing:** CRITICAL: All constant values in the \`arguments\` array of the SIR JSON (strings that are not variables) MUST be lowercase_snake_case (e.g., "earth", "gravity", "human", "force", "atom"). Variables MUST be ALL_CAPS (e.g., "X", "OBJECT"). If a natural language term like "FORCE" or "ATOM" refers to a type or category, it must be output as "force" or "atom" respectively when it's a constant argument. Do not use uppercase for constants unless they are proper Prolog variables.
+7.  **Specific Event Causality:** "Gravity makes apples fall."
+    Output: \`{"statementType": "fact", "fact": {"predicate": "causes_event", "arguments": ["gravity", "apples_fall"]}}\`
+8.  **Capability / General Rules based on Type:** "A force can cause acceleration."
+    Output: \`{"statementType": "rule", "rule": {"head": {"predicate": "can_cause_acceleration", "arguments": ["Thing"]}, "body": [{"predicate": "is_a", "arguments": ["Thing", "force"]}]}}\`
+9.  **General Composition/Definition of a Class:** "A molecule is composed of atoms."
+    Output: \`{"statementType": "fact", "fact": {"predicate": "general_composition_is", "arguments": ["molecule", "atoms"]}}\`
+
+- Ensure your JSON output is valid and strictly follows the structure described. Only output the JSON object.
+- Pay close attention to LEXICON SUMMARY for preferred predicate names and arities.
+- If significant assumptions are made, add a "translationNotes" field to the root of the JSON object.`,
+  user: `EXISTING FACTS (for context, do not translate these):
+\`\`\`prolog
+{{existingFacts}}
+\`\`\`
+
+ONTOLOGY RULES (for context, do not translate these):
+\`\`\`prolog
+{{ontologyRules}}
+\`\`\`
+
+LEXICON SUMMARY (preferred predicates and arities, use these if applicable, especially for predicate names and argument casing):
+\`\`\`
+{{lexiconSummary}}
+\`\`\`
+
+Based on all the context above, translate ONLY the following NEW natural language text into the SIR JSON format, strictly following the schema, casing rules for constants (lowercase) and variables (ALL CAPS) in arguments, and the key principles for translation:
+
+New Text: "{{naturalLanguageText}}"
+
+SIR JSON Output:`,
+};
