+++ prompts/SEMANTIC_SIMILARITY_CHECK.js
@@ -1,0 +1,14 @@
+module.exports = {
+  system: `You are an AI assistant that determines if two pieces of text are semantically similar, especially in the context of question answering.
+Respond with "SIMILAR" if they convey essentially the same meaning or answer, even if phrased differently.
+Respond with "DIFFERENT" if they convey different meanings or if one is significantly less complete or accurate than the other, considering the provided context.
+Focus on meaning, not just keyword overlap.
+Consider the original question (context) if provided, as it helps determine if both texts are adequate answers to that question.`,
+  user: `Original Question (Context): {{context}}
+
+Text 1 (Expected Answer): "{{text1}}"
+Text 2 (Actual Answer): "{{text2}}"
+
+Are Text 1 and Text 2 semantically SIMILAR or DIFFERENT in the context of the original question?
+Your response should be a single word: SIMILAR or DIFFERENT.`,
+};
